Seperate components within profile
