+++ src/containers/Jobs.jsx
@@ -23,7 +23,7 @@ import { FaFish } from "react-icons/fa";
 
 const iconSize = "1em";
 const Jobs = (props) => {
-  const { display, jobs } = props;
+  const { display, jobs, displayPanel } = props;
   const [displayJob, setDisplayJob] = useState(true);
   const [completion, setCompletion] = useState([0, 0]);
   const warMagicJobs = jobs.slice(0, 20);
@@ -72,7 +72,11 @@ const Jobs = (props) => {
 
       <div
         className={
-          "jobs__collection war-magic" + (displayJob ? "" : " disabled")
+          "jobs__collection" +
+          (displayPanel
+            ? " jobs__collection--single war-magic--single"
+            : " war-magic") +
+          (displayJob ? "" : " disabled")
         }
       >
         <JobHeader name="Tank" icon={tankIcon} />
@@ -94,7 +98,11 @@ const Jobs = (props) => {
       </div>
       <div
         className={
-          "jobs__collection hand-land" + (!displayJob ? "" : " disabled")
+          "jobs__collection" +
+          (displayPanel
+            ? " jobs__collection--single hand-land--single"
+            : " hand-land") +
+          (!displayJob ? "" : " disabled")
         }
       >
         <JobHeader name="Hand" icon={handIcon} />
