+++ src/styles/Character.css
@@ -3,15 +3,28 @@
     left: calc((100vw - 70rem) / 2);
     display: flex;
     flex-direction: column;
-    width: var(--content-width);
+    max-width: 70rem;
+    min-width: 70rem;
     gap: 2rem;
     margin: 8rem auto;
-    transition: left .5s, width .5s;
+    transition: left .5s, max-width .5s, min-width .5s;
+}
+
+.character__content {
+    display: flex;
+    gap: 2rem;
+}
+
+.character__tab {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
 }
 
 .character--side {
-    width: calc(100vw - 40rem);
-    left: 2rem;
+    max-width: calc(100vw - 41rem) !important; 
+    min-width: calc(100vw - 41rem) !important; 
+    left: 2rem !important;
 }
 
 .section {
@@ -23,26 +36,82 @@
     border-radius: .25rem;
     padding: 2rem;
     gap: 2rem;
-    background-color: var(--c-content-background);
+    background-color: var(--c-content-background-opaque);
     backdrop-filter: blur(5px);
     border: 1px solid var(--c-mid-background);
     box-shadow: 0 0 .8rem var(--c-shadow);
 }
 
-.navbar__overlay-panel-toggle {
+.overlay-panel__toggle {
     z-index: 4;
-    height: 4rem;
+    max-height: 3rem;
+    min-height: 3rem;
+    max-width: 3rem;
+    min-width: 3rem;
     position: fixed;
-    right: calc(32rem - 1px);
-    top: calc(50% - 2rem);
-    background-color: var(--c-content-background-opaque);
-    border: 1px solid var(--c-mid-background);
-    border-right: 1px solid var(--c-content-background);
+    right: 34rem;
+    top: calc(50% - 1.5rem);
     padding: 1rem;
-    border-radius: .5rem 0 0 .5rem;
+    color: var(--c-major-text);
+    background-color: var(--c-mid-background);
+    border: 1px solid var(--c-mid-background);
+    box-shadow: 0 .5rem 1rem var(--c-shadow);
+    border-radius: .25rem;
     transition: right .5s;
 }
 
-.navbar__overlay-panel-toggle--hide {
-    right: 0;
+.overlay-panel__toggle--hide {
+    box-shadow: none;
+    right: 2rem;
+    color: var(--c-minor-text);
+    background-color: var(--c-content-background-opaque);
+
+}
+
+.overlay-panel__avatar {
+    max-height: 2rem;
+    width: 2rem;
+    border-radius: 100%;
+    transition: transform .5s;
+}
+
+@media only screen and (max-width: 84rem) {
+
+    .character{
+        max-width: calc(100vw - 9rem);
+        min-width: calc(100vw - 9rem);
+        left: 2rem;
+    }
+
+}
+
+@media only screen and (max-width: 700px) {
+
+    .character__content{
+        flex-direction: column;
+    }
+
+    .character__tab {
+        flex-direction: row;
+        justify-content: center;
+    }
+
+    .character{
+        max-width: calc(100vw - 2rem);
+        min-width: calc(100vw - 2rem);
+        left: 1rem;
+    }
+    
+}
+
+@media only screen and (max-width: 73rem) {
+
+    .overlay-panel__toggle{
+        right: calc(100% - 5rem);
+    }
+
+    .overlay-panel__toggle--hide {
+        right: 2rem;
+    }
+
 }
