+++ src/styles/Equipment.css
@@ -1,0 +1,39 @@
+.equipment {
+    min-width: 28rem;
+    display: grid;
+    grid-template-columns: 3rem 20rem 3rem;
+    grid-template-rows: .5rem repeat(7, 3rem) .5rem;
+    grid-template-areas:
+        ".         portrait  .          "
+        "MainHand  portrait  OffHand    "
+        "Head      portrait  Earrings   "
+        "Body      portrait  Necklace   "
+        "Hands     portrait  Bracelets  "
+        "Legs      portrait  Ring1      "
+        "Feet      portrait  Ring2      "
+        ".         portrait  SoulCrystal"
+        ".         portrait  .          "
+    ;
+    margin: 0 auto;
+    gap: 1rem;
+}
+
+.equipment__portrait {
+    padding: .5rem;
+    display: flex;
+    align-items: flex-end;
+    grid-area: portrait;
+    background-position: center;
+    background-size: cover;
+    border-radius: .25rem;
+    box-shadow: 0 .5rem 1rem var(--c-shadow);
+}
+
+@media only screen and (max-width: 30rem) {
+
+    .equipment{
+        margin: 0;
+        grid-template-columns: 3rem 1fr 3rem;
+    }
+
+}
