+++ src/styles/Jobs.css
@@ -55,6 +55,63 @@
     ;
 }
 
+.jobs__collection--single {
+    grid-template-columns: 1fr;
+}
+
+.war-magic--single {
+    grid-template-areas: 
+    "Tank        "
+    "paladin     "
+    "warrior     "
+    "darkknight  "
+    "gunbreaker  "
+    ".           "
+    "Healer      "
+    "whitemage   "
+    "scholar     "
+    "astrologian "
+    "sage        "
+    ".           "
+    "Melee       "
+    "monk        "
+    "dragoon     "
+    "ninja       "     
+    "samurai     "   
+    "reaper      " 
+    ".           "
+    "Ranged      "   
+    "bard        "
+    " machinist  "
+    " dancer     "
+    ".           "
+    " Magic      "
+    " blackmage  "
+    "summoner    "
+    "redmage     "
+    "bluemage    "
+    ;
+}
+
+.hand-land--single {
+    grid-template-areas:
+    "Hand          "
+    "carpenter     "
+    "blacksmith    "
+    "armorer       "
+    "goldsmith     "
+    "leatherworker "
+    "weaver        "
+    "alchemist     "
+    "culinarian    "
+    ".             "
+    "Land          "
+    "miner         "
+    "botanist      "
+    "fisher        "
+    ;
+}
+
 @media only screen and (max-width: 700px) {
 
     .jobs__collection {
