+++ src/styles/Navbar.css
@@ -20,9 +20,3 @@
     height: 1.5rem;
     margin-right: 2rem;
 }
-
-.navbar__profile-icon {
-    height: 2rem;
-    width: 2rem;
-    margin: auto;
-}
