+++ src/styles/Quests.css
@@ -58,7 +58,7 @@
     
 }
 
-@media only screen and (max-width: 800px) {
+@media only screen and (max-width: 700px) {
 
     .quests__list {
         grid-template-columns: 1fr;
